requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
dataclasses-json==0.5.14
dnspython==2.4.1
exceptiongroup==1.1.2
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
langchain==0.0.257
langsmith==0.0.19
loguru==0.7.0
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
outcome==1.2.0
packaging==23.1
pinecone-client==2.2.2
pydantic==1.10.12
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.19
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
webdriver-manager==4.0.0
wsproto==1.2.0
yarl==1.9.2
